# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/shops": {
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "Retrieve a list of all shops",
                "operationId": "811428430320080e8bdf49e15e1352fb",
                "responses": {
                    "200": {
                        "description": "List of shops",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Shop"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Shops"
                ],
                "summary": "Create a new shop",
                "operationId": "eefcf4c24ec6cf6cd3f50d507d3afb02",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Shop"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Shop created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Shop"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create shop",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create shop"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shops/{id}": {
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "Retrieve the detail of a specific shop",
                "operationId": "edda6e2d8133e58ef707a839a9b2b694",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shop detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Shop"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Shop not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Shop not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Shops"
                ],
                "summary": "Update an existing shop",
                "operationId": "9c5760ebe2af4eeefd7adda2445d1305",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Shop"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Shop updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Shop"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Shop not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Shop not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to update shop",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to update shop"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Shops"
                ],
                "summary": "Delete a shop",
                "operationId": "64a490af502787f8c0d7e333d1ef482f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shop deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Shop deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Shop not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Shop not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete shop",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to delete shop"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Shop": {
                "title": "Shop",
                "description": "A store owned by a user that sells products.",
                "required": [
                    "user_id",
                    "name",
                    "contact_whatsapp",
                    "subdomain"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the shop",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "UUID of the user who owns the shop",
                        "type": "string",
                        "format": "uuid",
                        "example": "64cac925-f669-4621-9b5d-765af5a96dad"
                    },
                    "name": {
                        "description": "Name of the shop",
                        "type": "string",
                        "example": "Toko Fashion A"
                    },
                    "logo_url": {
                        "description": "URL to the shop logo",
                        "type": "string",
                        "example": "https://example.com/logo-a.png",
                        "nullable": true
                    },
                    "contact_whatsapp": {
                        "description": "WhatsApp contact number of the shop",
                        "type": "string",
                        "example": "081234567890"
                    },
                    "description": {
                        "description": "Detailed description of the shop",
                        "type": "string",
                        "example": "Trendy fashion store for young people.",
                        "nullable": true
                    },
                    "subdomain": {
                        "description": "Unique subdomain used for the shop's storefront",
                        "type": "string",
                        "example": "fashiona"
                    },
                    "created_at": {
                        "description": "Timestamp when the shop was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-01T08:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token to access secured endpoints",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}